% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calc_mineral_indices}
\alias{calc_mineral_indices}
\title{Calculate Diagnostic Spectral Mineral Indices}
\usage{
calc_mineral_indices(image)
}
\arguments{
\item{image}{A SpatRaster with calibrated B2, B4, B6, B7 bands.}
}
\value{
A SpatRaster containing Iron Oxide and Clay Mineral indices.
}
\description{
Calculate Diagnostic Spectral Mineral Indices
}
