% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{clean_image_noise}
\alias{clean_image_noise}
\title{Integrated Atmospheric and Surface Noise Masking}
\usage{
clean_image_noise(image, qa_layer, albedo_threshold = 0.5)
}
\arguments{
\item{image}{Multispectral SpatRaster (Reflectance).}

\item{qa_layer}{Landsat 9 QA_PIXEL layer for bitmasking.}

\item{albedo_threshold}{Threshold for bright surface removal (default 0.5).}
}
\description{
Integrated Atmospheric and Surface Noise Masking
}
