% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{export_mineral_maps}
\alias{export_mineral_maps}
\title{Batch Export Spectral Products to GeoTIFF}
\usage{
export_mineral_maps(mineral_raster, output_dir = "Results")
}
\arguments{
\item{mineral_raster}{SpatRaster to export.}

\item{output_dir}{Directory path for the output files.}
}
\description{
Batch Export Spectral Products to GeoTIFF
}
