% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{mask_env_noise}
\alias{mask_env_noise}
\title{Environmental Interference Masking (Vegetation and Water)}
\usage{
mask_env_noise(image, threshold = 0.3)
}
\arguments{
\item{image}{A SpatRaster containing B4 (Red) and B5 (NIR) layers.}

\item{threshold}{NDVI limit for masking vegetation (default 0.3).}
}
\description{
Environmental Interference Masking (Vegetation and Water)
}
