% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{plot_mineral_indices}
\alias{plot_mineral_indices}
\title{Plot Mineral Diagnostic Maps}
\usage{
plot_mineral_indices(mineral_raster)
}
\arguments{
\item{mineral_raster}{A SpatRaster from calc_mineral_indices.}
}
\description{
Plot Mineral Diagnostic Maps
}
