% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{plot_mineral_targets}
\alias{plot_mineral_targets}
\title{Generate Professional Exploration Target Plot}
\usage{
plot_mineral_targets(target_raster)
}
\arguments{
\item{target_raster}{A SpatRaster containing identified targets.}
}
\description{
Generate Professional Exploration Target Plot
}
