% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{scale_landsat_data}
\alias{scale_landsat_data}
\title{Calibrate Landsat 9 Digital Numbers (DN) to Reflectance}
\usage{
scale_landsat_data(image)
}
\arguments{
\item{image}{A SpatRaster object (Landsat 9 Collection 2 Level-1).}
}
\value{
A SpatRaster with reflectance values scaled between 0 and 1.
}
\description{
Calibrate Landsat 9 Digital Numbers (DN) to Reflectance
}
